one more fix on dtype for nxtomo

+++ ptychotools/utils/io.py
@@ -138,7 +138,7 @@ def create_nxtomo_file(filename, angles, shape, N, name='default', dtype=np.floa
     instrument.attrs['NX_class'] = 'NXinstrument'
     detector = instrument.create_group('detector')
     detector.attrs['NX_class'] = 'NXdetector'
-    detector.create_dataset('data', data=np.zeros((N,) + shape), dtype=dtype)
+    detector.create_dataset('data', data=np.zeros((N,) + shape, dtype=dtype), dtype=dtype)
     detector.create_dataset('image_key', data=np.zeros(N), dtype=int)
     sample = entry.create_group('sample')
     sample.attrs['NX_class'] = 'NXsample'
